成見 康宏 2018年1月15日
対応名：15期土曜勉強会_東京
コメント：SIJGM001Servise.javaをコミット

+++ src/main/java/jp/arise/sij/service/SIJGM001Servise.java
@@ -71,10 +71,14 @@ public class SIJGM001Servise {
 				Date date = new Date();
 			    long dateTimeTo2 = date.getTime();
 
-				//経過年数を算出
-			    long yearDiff = ( dateTimeTo1 - dateTimeTo2  ) / (1000 * 60 * 60 * 24 * 365);
+				//経過年数を算出し、int値に変換
+			    int yearDiff = (int)( dateTimeTo1 - dateTimeTo2  ) / (1000 * 60 * 60 * 24 * 365);
 
 				sijGm001Form.setPhase(list.get(i).getPhase());
+
+				//現段階で、COMGM003Dtoに使用路線を入力する項目がないため、エラー
+				//COMGM003Dtoにて作成してもらう必要がある。
+
 				sijGm001Form.setRoute(list.get(i).getRoute());
 				sijGm001Form.setPrice(list.get(i).getTanka());
 				sijgm001FormList.add(sijGm001Form);
